Run EPANET With Given Input File

Run EPANET With Given Input File

runEpanet(
  inpfile,
  returnOutput = FALSE,
  epanet.dir = getEpanetInstallationPath(),
  intern = FALSE,
  write.output = TRUE,
  ...,
  dbg = FALSE
)

Arguments

inpfile

full path to EPANET input file

returnOutput

if TRUE, the output is read from the generated output file and returned

epanet.dir

path to EPANET installation directory. Default: getEpanetInstallationPath()

intern

a logical, indicates whether to make the output of the command an R object.

write.output

if TRUE, EPANET will write a binary output file, else not

...

further arguments passed to readEpanetOutputFile, such as: read.prolog, read.energyUse, read.dynamicResults, read.epilog, see there.

dbg

if TRUE, debug messages are shown. Default: FALSE