Set REPORT Option in EPANET configuration

Set REPORT Option in EPANET configuration

Usage

setReportOptions(
  configuration,
  pagesize = 0,
  file = "",
  status = "NO",
  summary = "YES",
  messages = "YES",
  energy = "NO",
  nodes = "NONE",
  links = "NONE",
  variables = defaultReportVariables()
)

Arguments

configuration

EPANET configuration, representing an EPANET input file, as returned by readEpanetInputFile

pagesize

pagesize sets the number of lines written per page of the output report. The default is 0, meaning that no line limit per page is in effect.

file

file supplies the name of a file to which the output report will be written. If the file name contains spaces then it must be surrounded by double quotes. If not supplied then the Report file, as specified in the second parameter of the ENopen (or ENepanet) function will be used.

status

status determines whether hydraulic status messages are written to the Report file. If YES is selected the messages will identify those network components that change status during each time step of the simulation. If FULL is selected, then convergence information will also be included from each trial of each hydraulic analysis. This level of detail is only useful for de-bugging networks that become hydraulically unbalanced. The default is NO.

summary

summary determines whether a summary table of number of network components and key analysis options is generated. The default is YES.

messages

messages determines whether error and warning messages generated during a hydraulic/water quality analysis are written to the Report file. The default is YES.

energy

energy determines if a table reporting average energy usage and cost for each pump is provided. The default is NO.

nodes

nodes identifies which nodes will be reported on. You can either list individual node ID labels or use the keywords NONE or ALL. Additional NODES lines can be used to continue the list. The default is NONE.

links

links identifies which links will be reported on. You can either list individual link ID labels or use the keywords NONE or ALL. Additional LINKS lines can be used to continue the list. The default is NONE.

variables

list of report variables as defined by reportVariable, e.g. list(reportVariable(name = "VELOCITY", above = 3.0, precision = 4), reportVariable(name = "F-FACTOR", precision = 4)). Default: defaultReportVariables